osc preview....

+++ update_app/update_app.cpp
@@ -865,10 +865,14 @@ static int pro2Updatecheck(const char* pUpdateFileDir)
 
 #define INSTALL_SAMBA_CMD	"/usr/local/bin/install_samba.sh"
 
+
 static void installVm()
 {
-	mkdir("/swap", 0766);
-	system("dd if=/dev/zero of=/swap/sfile bs=1024 count=4000000");
+	if (access("/swap/sfile", F_OK) != 0) {
+		mkdir("/swap", 0766);
+		system("dd if=/dev/zero of=/swap/sfile bs=1024 count=4000000");
+	}
+	
 	system("mkswap /swap/sfile");
 	system("swapon /swap/sfile");
 	system("cp /usr/local/bin/fstab /etc/fstab");
@@ -986,6 +990,8 @@ static int start_update_app(const char* pUpdatePackagePath, bool bMode)
 					chmod(INSTALL_SAMBA_CMD, 0766);
 					installSamba(INSTALL_SAMBA_CMD);
 				}
+
+				installVm();
 			}
 		}
